requirements.txt
pyserial
hotqueue
future